+++ cadastral_automation/__init__.py
@@ -1,0 +1,12 @@
+"""
+Cadastral Automation Plugin
+"""
+
+def classFactory(iface):
+    """Load CadastralAutomation class from file cadastral_automation.
+    
+    Args:
+        iface: A QGIS interface instance.
+    """
+    from .cadastral_automation import CadastralAutomation
+    return CadastralAutomation(iface)